+++ app/src/main/java/com/example/internetapi/repository/AccountRepository.kt
@@ -8,8 +8,8 @@ class AccountRepository @Inject constructor(
     private val apiHelper: ApiHelper
 ) {
     suspend fun getAccounts() = apiHelper.getAccounts()
-    suspend fun getAccountInvoices(accountId: Long) = apiHelper.getAccountInvoices(accountId)
+    suspend fun getAccountInvoices(accountId: Int) = apiHelper.getAccountInvoices(accountId)
     suspend fun getInvoiceDetails(invoiceId: Long) = apiHelper.getInvoiceDetails(invoiceId)
-    suspend fun getAccountIncome(accountId: Long) = apiHelper.getAccountIncome(accountId)
+    suspend fun getAccountIncome(accountId: Int) = apiHelper.getAccountIncome(accountId)
     suspend fun updateAccount(accountId: Int, updateAccountRequest: UpdateAccountRequest) = apiHelper.updateAccount(accountId, updateAccountRequest)
 }
